Added writeStringBuffer and corresponding connect method
Copies the connect method, with the change being to have a death payload as a buffer and buffer length instead of just a const char. This makes it possible for serialized death payload (eg. google protobuffers) that may contain a 0 value byte (const char will stop at a 0 as this considered null character)